+++ django_field_cryptography/tests/test_fields.py
@@ -3,7 +3,7 @@
 from django.test import TestCase
 
 from .. import fields
-from . import factories
+from . import factories, models
 
 
 class TestField(TestCase):
@@ -31,5 +31,9 @@ class TestField(TestCase):
 class TestModel(TestCase):
     def test_value(self):
         value = 'Encrypt me'
-        dummy_model = factories.DummyModelFactory.create(aes_field=value)
+        factories.DummyModelFactory.create(aes_field=value)
+
+        # requerying the model allow us to call `to_python` to test if the
+        # decryption is handled properly.
+        dummy_model = models.DummyModel.objects.get()
         self.assertEqual(dummy_model.aes_field, value)